requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
statsmodels>=0.14.0
scikit-learn>=1.3.0
openai>=1.0.0
python-dotenv>=1.0.0
swarm @ git+https://github.com/openai/swarm.git
